requirements.txt
fastapi~=0.115.13
uvicorn~=0.34.3
pandas~=2.3.0
numpy~=1.26.0
statsmodels~=0.14.4
requests~=2.32.4
matplotlib~=3.9.4
scipy~=1.13.1
seaborn~=0.13.2
pillow~=11.2.1
starlette~=0.46.2
pydantic~=2.11.7
python-dateutil~=2.9.0.post0
python-multipart
networkx~=3.2.1
PyYAML~=6.0.2
rasterio~=1.4.3
shapely~=2.0.7
geopandas~=1.0.1